First steps
In order to analyse our results we need to first create a script. I already made a script starting today where I ended up with all this code. The code contains comments that guide the reader through the process, although not fully complete yet. This will not be the final project and is not even a preliminary version, just a first draft of the rough analysis. I've also added the excel sheet with the data to the github repository.