Streamlined the Code
To start, I've made some changes to the main file. I've cleaned up some of the lines that were obsolete and not contributing to the code OR werent working anymore. Next I've changed the plotting function from a line/scatter plot to a boxplot and also incorporated the save.pdf function. I've restructured the code into three main parts: preparatory analysis, plotting graphs and statistical analysis. I've shifted the analysis from max and min to the mean temperature and RH. Lastly I've added the anova analysis for the temperature variable aswell.

I've generated the two plots and added them to the output file. Ive also removed the "Gene expression" plot from the file. This was an empty file that was serving as a test and placeholder for the now current plots